Change `list` to `manage` & add init function
